non cgo build
(cherry picked from commit cffe53fa0eeee0a9735ac2dd6d419d9d1d25eccf)
