+++ x/wasm/client/cli/query.go
@@ -10,7 +10,6 @@ import (
 	"os"
 	"strconv"
 
-	wasmvm "github.com/CosmWasm/wasmvm"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -45,27 +44,6 @@ func GetQueryCmd() *cobra.Command {
 		GetCmdListContractsByCreator(),
 	)
 	return queryCmd
-}
-
-// GetCmdLibVersion gets current libwasmvm version.
-func GetCmdLibVersion() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:     "libwasmvm-version",
-		Short:   "Get libwasmvm version",
-		Long:    "Get libwasmvm version",
-		Aliases: []string{"lib-version"},
-		Args:    cobra.ExactArgs(0),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			version, err := wasmvm.LibwasmvmVersion()
-			if err != nil {
-				return fmt.Errorf("error retrieving libwasmvm version: %w", err)
-			}
-			fmt.Println(version)
-			return nil
-		},
-		SilenceUsage: true,
-	}
-	return cmd
 }
 
 // GetCmdBuildAddress build a contract address
