+++ x/wasm/libwasm_version_no_cgo.go
@@ -1,0 +1,11 @@
+//go:build !cgo
+
+package wasm
+
+import (
+	"github.com/spf13/cobra"
+)
+
+func checkLibwasmVersion(cmd *cobra.Command, args []string) error {
+	panic("not implemented, please build with cgo enabled")
+}
