+++ x/wasm/module.go
@@ -3,12 +3,9 @@ package wasm
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"math/rand"
 	"runtime/debug"
-	"strings"
 
-	wasmvm "github.com/CosmWasm/wasmvm"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
 	cdctypes "github.com/cosmos/cosmos-sdk/codec/types"
@@ -278,21 +275,6 @@ func getExpectedLibwasmVersion() string {
 		return d.Version
 	}
 	return ""
-}
-
-func checkLibwasmVersion(cmd *cobra.Command, args []string) error {
-	wasmVersion, err := wasmvm.LibwasmvmVersion()
-	if err != nil {
-		return fmt.Errorf("unable to retrieve libwasmversion %w", err)
-	}
-	wasmExpectedVersion := getExpectedLibwasmVersion()
-	if wasmExpectedVersion == "" {
-		return fmt.Errorf("wasmvm module not exist")
-	}
-	if !strings.Contains(wasmExpectedVersion, wasmVersion) {
-		return fmt.Errorf("libwasmversion mismatch. got: %s; expected: %s", wasmVersion, wasmExpectedVersion)
-	}
-	return nil
 }
 
 type preRunFn func(cmd *cobra.Command, args []string) error
